Delete old-avatar in cloudinary and update session after update avatar

+++ adminLTE/controller/admin/profile.controller.js
@@ -69,24 +69,53 @@ const renderEditProfile = async (req, res, next) => {
 };
 
 const updateInfo = async (req, res, next) => {
+    const { user } = req.session;
     const { userID } = req.params;
     const { path } = req.file;
     let imageAfterUpload;
+    console.log(`userID: ${userID}`);
+    if (user.id !== userID && user.role !== 'admin') {
+        req.flash(
+            'blankMessage',
+            'Dont allow access edit infomation of other users'
+        );
+        return res.redirect('/admin/blank-message');
+    }
 
     try {
+        // ? upload image in local to cloudinary
         imageAfterUpload = await cloudinary.uploader.upload(path, {
             tags: 'avatar',
             folder: 'VyXinhGais-Blog/avatar',
         });
 
+        // ? delete old-avatar in cloud
+        const oldDataOfUser = await knex('users')
+            .select()
+            .where({ id: userID })
+            .first();
+        const oldPulicId = oldDataOfUser.avatar.match(
+            /VyXinhGais-Blog\/avatar\/\w+/
+        )[0];
+        console.log('oldPublicID: ', oldPulicId);
+        const result = await cloudinary.uploader.destroy(oldPulicId, {
+            invalidate: true,
+        });
+        console.log(result);
+
+        // ? update new avatar and delete file in local
         await knex('users')
             .where({ id: userID })
             .update({
                 avatar: imageAfterUpload.url,
             });
-
-        console.log(`* ${imageAfterUpload.url}`);
         fs.unlinkSync(path);
+
+        console.log(`* ${imageAfterUpload.public_id}`);
+        console.log(`* ${imageAfterUpload.url}`);
+        if (req.session.user.id === userID) {
+            req.session.user.avatar = imageAfterUpload.url;
+        }
 
         req.flash('message', {
             status: 'success',
